Finish README
Hello Everyone! I'm Galaxy a student of Nueva Ecija University of Science and Technology Graduate School. I really want to be a professional programmer someday. The Languages that I am currently focusing on are java, python and C++ and the Framework that catch my interest for  Web programming are Django and Spring. I have high hopes for the professional members here in github, that I will learn a lot of new knowledge when it comes to programming. Thank you and God Bless everyone!